// src/src/blockchain-indexer/rpc/types/BlockStats.ts
export interface BlockStats {
    avgfee: number;
    avgfeerate: number;
    avgtxsize: number;
    blockhash: string;
    feerate_percentiles: [number, number, number, number, number];
    height: number;
    ins: number;
    maxfee: number;
    maxfeerate: number;
    maxtxsize: number;
    medianfee: number;
    mediantime: number;
    mediantxsize: number;
    minfee: number;
    minfeerate: number;
    mintxsize: number;
    outs: number;
    subsidy: number;
    swtotal_size: number;
    swtotal_weight: number;
    swtxs: number;
    time: number;
    total_out: number;
    total_size: number;
    total_weight: number;
    totalfee: number;
    txs: number;
    utxo_increase: number;
    utxo_size_inc: number;
}